split worker and rdd into one class per file to easy livy compilation
